refactor(employee-components): reorganize component imports and structure
