+++ features/employee/index.ts
@@ -1,3 +1,3 @@
-export { EmployeeCard } from './employee-card';
-export { CurrentProjectsCard } from './current-projects-card';
-export { EmployeeProjectsCell } from './employee-projects-cell';
+export { EmployeeCard } from './components/employee-card';
+export { CurrentProjectsCard } from './components/current-projects-card';
+export { EmployeeProjectsCell } from './components/employee-projects-cell';
